requirements.txt
numpy
pandas
torch
torchvision
pillow
opencv-python
psutil
scipy
matplotlib